chore: flex refac first try
