+++ internal/commands/list.go
@@ -1,0 +1,36 @@
+package commands
+
+import "github.com/bwmarrin/discordgo"
+
+var List = []*discordgo.ApplicationCommand{
+	{
+		Name:        string(EPivo),
+		Description: "Ranking by the times we met",
+		Options: []*discordgo.ApplicationCommandOption{
+			{
+				Type:        discordgo.ApplicationCommandOptionUser,
+				Name:        "user-option",
+				Description: "User option",
+				Required:    true,
+			},
+			{
+				Type:        discordgo.ApplicationCommandOptionRole,
+				Name:        "role-option",
+				Description: "Role option",
+				Required:    true,
+			},
+		},
+	},
+	{
+		Name:        string(ELoh),
+		Description: "Лох",
+		Options: []*discordgo.ApplicationCommandOption{
+			{
+				Type:        discordgo.ApplicationCommandOptionUser,
+				Name:        "user-option",
+				Description: "User option",
+				Required:    true,
+			},
+		},
+	},
+}
