+++ internal/commands/types.go
@@ -1,0 +1,8 @@
+package commands
+
+type SvineBotCommand string
+
+const (
+	EPivo SvineBotCommand = "pivo"
+	ELoh  SvineBotCommand = "loh"
+)
