+++ magefiles/main.go
@@ -3,13 +3,13 @@
 package main
 
 import (
-    "github.com/magefile/mage/sh"
+	"github.com/magefile/mage/sh"
 )
 
 func Build() error {
-    return sh.Run("go", "build", "-o", "tmp", "./...")
+	return sh.Run("go", "build", "-o", "tmp", "./...")
 }
 
 func BuildLive() error {
-    return sh.RunV("air", "*> ./tmp/run-verbose.log")
+	return sh.RunV("air", "*> ./tmp/run-verbose.log")
 }
